requirements.txt
asgiref==3.2.10
Django==3.1.2
mysql-connector-python==8.0.21
mysqlclient==2.0.1
numpy==1.19.2
pandas==1.1.3
protobuf==3.13.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
